refactor: move `formatCpp` function to `generate/format.ts` file
